backend/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
pypdf==4.3.1
pillow==10.4.0
python-multipart==0.0.9
